requirements.txt
numpy>=1.21.0
matplotlib>=3.4.0
scikit-learn>=1.0.0
pandas>=1.3.0
seaborn>=0.11.0
tensorflow>=2.8.0
keras>=2.8.0
pillow>=8.0.0
tqdm>=4.62.0
librosa>=0.9.0
soundfile>=0.11.0
audioread>=2.1.9
umap-learn>=0.5.3
